requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
python-docx
PyMuPDF
requests
dateparser
